Add console1984
We want a way of tracking console sessions. console1984 provides this
with sensible defaults.

It is configured to request a name when starting a session, along with a
purpose for the session.

All the commands entered are stored in the new tables and we will
introduce a companion library, audits1984, in a future change to be able
to review them.