Add audits1984
We have added tracking to our console sessions and now are introducing
audits1984 as a way to audit those sessions.

This gives us an interface to be able to review all the console
sessions and either approve or flag them for issues.

It uses the Staff authentication system and any Staff member has access
currently.

Flagging a session as inappropriate only sets the field in the Find
database. We will want to tie it into any existing procedure we have for
inappropriate access/use of data. That is beyond the scope of this
change.