Add a remote terminal in the support interface
We want a way to access the Rails console in production but in an
auditable way.

We added console1984 and audits1984 to protect encrypted data and also
record the commands that are run in a console session.

The next step is to provide an interface to the console via the support
interface, that uses the staff authentication system to control access.

I explored using the web-console library that is included in the default
Rails install. It isn't suited to providing remote access to the
terminal in a production environment. In fact, the docs specifically
recommend against it.

It is very coupled to being triggered when an exception occurs. This
results in needing to monkey patch the library to prevent this happening
in production.

As an alternative approach, I used the rvt library instead. It is a
Rails engine that manages setting up xterm.js to run the Rails console.

Out of the box, it renders a fullscreen terminal and again is set up to
only run in development.

In order to secure access to it, I extracted the view code from the
engine and embedded it directly in a view we control that is protected
by Devise.

The result is that we have a section in the support interface that
provides authenticated access to a Rails console, running in that
environment.

I wasn't able to successfully pass the details of the currently
authenticated staff member to the console auditing library
(console1984).

For now, we need to rely on people entering a suitable value at the name
prompt when using the console.