Override Rails convention for credentials path
The Rails convention for the location of the credentials file relies on
the RAILS_ENV matching a file in `config/credentials`.

We can't use that convention because all our hosted applications run
with `RAILS_ENV=production`.

As a short-term alternative, we can use `HOSTING_ENVIRONMENT_NAME` to
ensure that only the actual production app uses the production
encryption key.

Every other environment will use the key we use in development.