Throw an exception when finding a ticket fails
When requesting a ticket from Zendesk, any error messages are silent and
instead we get a `NoMethodError` exception due to calling a method on an
object we expect to be a ticket.

To help diagnose these issues more easily in the future, we can use the
`find!` method on the Zendesk client to raise an exception about why the
API call failed.