Enable querying by email
The email field on a TrnRequest is encrypted using a non-deterministic
key, which makes querying it impossible.

Changing this is a 2-step process.

Firstly, we need to define the attribute as requiring a deterministic
key.

On its own, this would prevent us being able to access the email data
because Rails wouldn't be able to decrypt the field anymore as it is
expecting a derterministic key.

We can solve that by using the `previous` configuration option. This
allows us to specific the previous encryption scheme used on that field.

There is another configuration option that must be set as well, which is
`fixed: false`. This ensures that the encryption scheme can handle the
non-deterministic original values.

We then can re-encrypt all the existing data using a data migration.

The second part of this change will happen in a follow-up PR. We need to
remove the extra configuration options once the data migration has run
successfully. This will ensure that the email field is encrypted with a
deterministic key correctly for all future records.

Ideally, we'd deploy both sets of changes sequentially, at a time when
we don't expect there to be any new records created. To be certain that
this can't happen, I recommend that we disable the service during these
deployments.

Also, for local development, the same process needs to be followed if
you wish to preserve data that was encrypted with the previous scheme.

Alternatively, it is safe to reset the DB and start with fresh data.