Enable querying by email
The email field on a TrnRequest is encrypted using a non-deterministic
key, which makes querying it impossible.

This is the second step in a 3 step process to switch to a deterministic
key.

On its own, this would prevent us being able to access the email data
because Rails wouldn't be able to decrypt the field anymore as it is
expecting a deterministic key.

We can solve that by using the `previous` configuration option. This
allows us to specify the previous encryption scheme used on that field.

There is another configuration option that must be set as well, which is
`fixed: false`. This ensures that the encryption scheme can handle the
non-deterministic original values.

We then can re-encrypt all the existing data using a data migration.

The final part of this change will happen in a follow-up PR. We need to
remove the extra configuration options once the data migration has run
successfully. This will ensure that the email field is encrypted with a
deterministic key correctly for all future records.