Decrypt all data
We want to encrypt the email field using a deterministic encryption
scheme. To get there, we need to re-encrypt it.

I propose doing this in 3 steps (structured in 3 separate PRs).

First, we decrypt all the TrnRequest data and re-introduce support for
unencrypted data to ensure the app continues to run as normal.

The second step is to configure the email field to use deterministic
encryption and then encrypt all the fields again.

Finally, the third step is to remove support for unencrypted data, so we
don't accidentally leave any data exposed in plaintext.

This process allows us to make changes to the encrypted fields without
introducing any downtime to the service.

The cost of this is that for the duration of time between deploying
steps 1 and 3, the data in the service will be unencrypted and viewable
as plaintext.