Remove un-needed encryption configuration
During the transition phase of enabling querying by email, we added some
configuration options to ActiveRecord encryption to support the
conversion.

We no longer need these settings and can safely remove them.

The one setting that is remaining is `store_key_references`. Once
enabled, this is requires more effort to remove.

A reference to the encryption key is stored in the ciphertext headers,
making the lookup more performant, in exchange for database space.