Link the service name to the service URL
Currently, the service name is linked to the Find root path. This could
cause confusion for users who came to find their TRN from a different
service.

We accept a client_url in the params of the identity flow and can use
this, if present, to link the service name to the calling service
correctly.