Use Continue rather than Submit for the identity flow
When a user has arrived via the identity flow, the Submit button on the
check answers page is misleading, because the flow doesn't end there.

A more appropriate term is `Continue`.

This change ensures that when a TrnRequest is generated via the identity
flow, that we display the appropriate term.