Make DeleteOldZendeskTicketsJob work with >100 tickets
We built this job but purposefully added in a guard clause / raise to break if it's run with more than 100 tickets.

The ZendeskService.find_closed_tickets_from_6_months_ago can return a maximum of 100 tickets and is a paginated request. destroy_tickets! can only operate with a maximum of 100 tickets.

Run the job recursively with more than 100 tickets