Add a Slack alert for Zendesk deletions
In a previous change, we scheduled the automatic deletion of old Zendesk
tickets to happen weekly.

We want to have some visibility of this process during the first few
weeks of it running to give us confidence that it is working as
expected.

A proposal is that a Slack alert 24 hours before the scheduled deletion
runs will give us sufficient notice to address any irregularities in the
expected count of tickets scheduled to be deleted.