Revert "Use /health/all endpoint for healthchecks"
This reverts commit 7a0ce4b4a5445147ea1175358b81dc5fa778bbec.

This endpoint is used by the PaaS for more than just checking the health
of the container during a deployment; it's also pinged during regular
operation. What this means is that if one of Find's dependencies goes
down, like Zendesk or Notify, PaaS will think the container is unealthy,
and potentially kill it.