Use the count method provided by ZendeskAPI
The ZendeskAPI library includes a method, `count` on the
`ZendeskAPI::Collection` object.

We are not using this object but instead using `fetch` and working with
the JSON response from the API.

Rather than read the count value from the API response, it seems simpler
to lean on the interface provided by the library.

The change for this is small. In fact, most of the changes are in our
test code.