Add DownloadHelpers for system specs
Cuprite doesn't handle downloads very elegantly out of the box. It
doesn't offer a documented way to wait for a `network_traffic.last` to
complete.

This adds a `DownloadHelpers` module taken from a GitHub comment that
effectively sleeps in a loop until downloads get written to disk,
followed by reading the file and providing the content for assertions.