Don't run DeleteOldZendeskTicketsJobs on non-prod
This job should only run on the production environment, because it
destroys Zendesk tickets and populates the database with the delete
requests. If it can run on other envs, we run the risk of having
multiple envs as the source of truth for previously deleted tickets,
which has unfortunately happened (preprod has 131 `ZendeskDeleteRequest`s).