Add system spec for support interface teacher page
We're implementing a teacher show page in the support interface. This
page will eventually allow support users to handle various aspects of
the teacher's identity. Initially we're going to add a skeleton page
with stubbed out data.

As a first step, add a system spec outlining interaction with this page.