Allow authenticating via Identity
We want to allow staff to authenticate via Identity to get access to the
support interface.

Identity has an OAuth2 server that we have a client for. Using Omniauth,
we are able to add Identity as an authentication provider.

As a first iteration, I opted to keep the data model simple and trust
the data returned by Identity. This means that we either match on email
for existing Staff records or create them when there is no match.

We have to override the redirect_uri parameter for the access token call.
OmniAuth assumes that the server will be able to handle a redirect_uri
containing querystring params. However, the implementation we are
working with needs them to be an exact match, so no querystring.

As a result, I redefined the method that retrieves the access token.