Move recursion to the tail end of the job
DeleteOldZendeskTicketsJob checks if there are more than 100 tickets to
delete. If there are, it queues another job, then tries to delete them.

The problem is that if the delete call fails, it also queues another
job. So you can queue 2 jobs from 1 job. Then those 2 jobs can queue up
to 4 jobs. And so on.

This is a temporary fix to move the recursion to the end of the call. A
better fix is to rewrite the job to use pagination/fetch the full list
of Zendesk tickets instead of relying on recursion.