Use a RESTful controller for CSV exports
We originally built the CSV exports using a custom action that responded
to a GET request.

I have a preference for following Rails conventions of using RESTful
resources. A CSV export action should be a resource of its own and also,
it should use a POST verb to represent that we are creating a resource.