Remove duplicates from the exported CSV
The list of ZendeskDeleteRequests that we return as a CSV includes
duplicate ticket IDs.

A future change should be to address this properly with validations on
the model and an index and constraint on the database column. This would
require a strategy to manage the duplicate delete requests.

Rather than do that now, I opted to change the scope we use for
populating the CSV and de-dupe via a query.

I decided to make the query customisable and default to the no-dupes
version to maintain the existing interface.