Change the format of the returned CSV fields
We have a request to change the formatting of the date column and also
the "No action required" column.

The timezone info is extraneous because everyone using this CSV is in
the UK. We can default to displaying all datetimes in that timezone.

Also, we want to make the No action required column easier for people to
parse by ensuring that the value "t" is represented by "No action
required" and everything else is ignored.

Finally, there is a request to make the enquiry_type column human
friendly.