Allow nil Zendesk ticket group name
We have an exception in production where a Zendesk ticket is missing a
group value.

This causes the app to crash because it expects that the group value
will always be present.

As an interim fix, I opted to make it explicit that this value can be
nil.

Once we ship this change, I can investigate the issue again via the
support interface as I don't have access to production data any other
way.