Remove repetition in Identity system specs
In the 'no TRN match' scenario, there are several example groups that
repeat the same basic set of steps in order to assert various things
about this user journey. The example group descriptions don't need to be
this granular in a system test - the individual step methods provide the
detail.

Make the following changes:

- Remove all but one example group
- Give it an overarching description
- Consolidate the various 'then' steps into this example group