Configure sidekiq critical and low priority queues
Sidekiq allows us to specify multiple queue names, and they are given
priority in the order they are declared in.

This sets up a Critical priority queue which is used for Zendesk ticket
creation and for Account unlocking functionality, and a Low priority
queue which is used for the Zendesk ticket deletion job. All other jobs
by default will execute in the medium priority default queue.