Remove redirect to DQT records form
We used to redirect from the users (trn) update action to the dqt edit action which exposed the TRN in the redirect URL.
We stopped exposing the trn by storing it in the session but this would
cause side-effects with multiple forms being open across tabs as the
same trn in session would be used.
Render the dqt_records edit action instead, this hides the TRN but requires a bit of duplication of API calls in the update action.