Amend logic in EnforceQuestionOrder
Rather than just checking if the previous question has been answered,
check all prior questions. The existing implementation of
'#next_question_path' will then redirect to the first unanswered
question in the flow.

This fixes an issue in the Find a TRN journey. With the addition of
preferred name, the early question flow is:

- email
- name
- preferred name
- date of birth

A Find journey that submits the email and then skips ahead to date of
birth will check the preferred name flow. Because this doesn't need
answering, the user is sent on to the date of birth page rather than
being redirected to name.

This code change ensures a redirect to the earliest unanswered question
takes place.