Rewrite Zendesk deletion job
The old job:

- Calls the Search endpoint
- Takes the first 100 results (the max per page)
- Attempts to delete those
- Schedules another job to run 5 minutes later

This would recursively delete all the tickets. However, because the
`destroy_many!` endpoint on the ZendeskAPI is async, and doesn't
immediately destroy the tickets, it would sometimes take multiple
attempts of attempting to destroy the same 100 tickets over and over
before it would eventually get through.

The advantage of this old approach is that it was blunt and worked
around the limitations in the Zendesk API [^0]. The search endpoint only
returns a maximum of 1000 results, paginated at 100 items per page. To
get to the 1001st ticket, you have to modify the query and add more
filters, like clamping the dates, to reduce the total number of tickets.

The disadvantage was that it was inefficient, and it would take up to a
week to delete e.g. 1700 tickets. We also couldn't run this job more
than once a week, because it wouldn't make a difference.

The new job:

- Calls the Search endpoint
- Walks through all 10 pages
- Schedules a job for each page, staggered at 5 minutes, to delete the
  tickets

This can be run daily and should hopefully be more consistent.

The disadvantage to this new approach is that if on a particular day
there are more than 1000 tickets to delete, it won't even attempt to do
so. So eventually, this approach might break if the service scales too
much, but that seems unlikely at the moment.

[^0]: https://developer.zendesk.com/api-reference/ticketing/ticket-management/search/