Add paging params to the Identity users endpoint
The Identity API endpoint for users has recently added support for
pagination.

We can now pass query params to page the results via the API rather than
fetching all users and paging in memory.

This should result in a more performant initial load for the users view
in the support interface.

The new params are `pageNumber` and `pageSize`.

The default page size in the API is 25 but given our interface has an
expectation on paging 100 records at a time, I supplied that as our
default.