Remove session reset from Identity redirect
Clearing a session on redirect to Identity may cause issues when users
use their back button.

The change in #689 ensures that we clear the session whenever we receive
a new request from GAI.

I believe the reset on redirect was intended to perform the same
function so we can safely remove it.