Allow users with active sanctions to continue to Identity
There have been cases where users in the identity flow have a TRN that
has active sanctions.

For the Get an Identity journey, having active sanctions isn't a reason
to prevent us from saving the TRN on their DQT record and sending them
to Zendesk.

The approach we settled on was to allow users on the identity journey to
continue back to GAI regardless of the presence of active sanctions.