Ensure the TRN export has a recipient
We need to ensure that the `TrnExportMailer` has a recipient to send
the email to.

I opted to use the Rails credentials file to store this as it gives us
greater control over the value without needing extra permissions to set
environment variables.

This value doesn't feel like an environment specific one. It seems to
strange to have to declare this kind of value in a Dockerfile for
instance.