Add a CSV column representing account unlocking
The CSV of TRNs that we export is missing a data on whether the account
was also unlocked. This is a key part of the analysis the team are
currently wanting to do.

This change adds a `First Unlocked At` column to represent when the
unlock event was created for that particular request.