Disable deleting Zendesk tickets
We are about to switch Zendesk instance and want to avoid deleting
tickets that are un-related to Find.

This change removes the schedule for deleting tickets but leaves the
functionality available in case we want to re-enable it.

There is a Trello ticket for investigating how we want to manage our
data retention policy on this alternative Zendesk instance.