Fix ZendeskDeleteRequest bug following Ruby upgrade
The method ZendeskDeleteRequest.from_csv returns nil even after
successfully creating a record. This might be because of a change in
behaviour in `CSV.foreach`.

This causes ZendeskDeleteRequestImportForm#save to return nil even when
a valid save has taken place. Fix this by explicitly returning true at
the end of the method.