Update webhook env var in SlackClient
This env var hasn't been set for the production app since the AKS
migration, and is needed by SlackClient for reporting performance alerts
and Zendesk health to Slack.

Secrets for this service are now split into infra and app lists. The
infra list contains a Slack webhook URL that is used when database
backups fail. We need that same URL in the app list.

In addition to adding SLACK_WEBHOOK to the app secrets via the Azure
Portal, this PR simply aligns the name of the env var between the two
places in the codebase it is currently used.